requirements.txt
click
paho-mqtt
loguru
icecream
shapely
isort
haversine
